Convert wait_event_info to 64-t bits, expose lower 32-bit as new wait_event_arg
Implement some very basic uses of wait_event_arg for the following wait events:
- Timeout/PgSleep
- IPC/RecoveryConflictTablespace

Due to (Win32-based) compiler limitations, we need to resign from use of enums
as they do not always allow use of 64-bit integers.

Author: Jakub Wartak <jakub.wartak@enterprisedb.com>
Discussion: https://www.postgresql.org/message-id/flat/CAKZiRmyKcTaeSGzMYDN6aRR-BwYGPeZbzDRKvGkJhxAghfb4LQ%40mail.gmail.com

+++ contrib/dblink/dblink.c
@@ -143,9 +143,9 @@ static remoteConn *pconn = NULL;
 static HTAB *remoteConnHash = NULL;
 
 /* custom wait event values, retrieved from shared memory */
-static uint32 dblink_we_connect = 0;
-static uint32 dblink_we_get_conn = 0;
-static uint32 dblink_we_get_result = 0;
+static uint64 dblink_we_connect = 0;
+static uint64 dblink_we_get_conn = 0;
+static uint64 dblink_we_get_result = 0;
 
 /*
  *	Following is hash that holds multiple remote connections.
